move optional scalars test to own file:
