+++ tests/optional_scalars_test.h
@@ -1,0 +1,13 @@
+#ifndef TESTS_OPTIONAL_SCALARS_TEST_H
+#define TESTS_OPTIONAL_SCALARS_TEST_H
+
+
+namespace flatbuffers {
+namespace tests {
+
+void OptionalScalarsTest();
+
+}  // namespace tests
+}  // namespace flatbuffers
+
+#endif
